Removed uint64_t from scl_graph and modyfied sort_algos

+++ src/include/scl_graph.h
@@ -37,7 +37,7 @@
  * 
  */
 typedef struct graph_link_s {
-    uint64_t vertex;                                        /* Vertex number that links with the original vertex */
+    size_t vertex;                                          /* Vertex number that links with the original vertex */
     long double edge_len;                                   /* Length of the edge starting from original vertex to number vartex */
     struct graph_link_s *next;                              /* Pointer to the next link or edge from the graph */
 } graph_link_t;
@@ -65,32 +65,32 @@ typedef struct graph_s {
 graph_t*            create_graph                            (size_t number_of_vertexes);
 scl_error_t         free_graph                              (graph_t * const __restrict__ gr);
 
-scl_error_t         graph_insert_edge                       (const graph_t * const __restrict__ gr, uint64_t first_vertex, uint64_t second_vertex, long double edge_len);
+scl_error_t         graph_insert_edge                       (const graph_t * const __restrict__ gr, size_t first_vertex, size_t second_vertex, long double edge_len);
 scl_error_t         graph_insert_vertices                   (graph_t * const __restrict__ gr, size_t new_vertices);
 graph_t*            create_transpose_graph                  (const graph_t * const __restrict__ gr);
 scl_error_t         graph_print                             (const graph_t * const __restrict__ gr, const uint8_t ** const data_arr);
 
-scl_error_t         graph_delete_edge                       (const graph_t * const __restrict__ gr, uint64_t first_vertex, uint64_t second_vertex);
-scl_error_t         graph_delete_all_edges                  (const graph_t * const __restrict__ gr, uint64_t first_vertex, uint64_t second_vertex);
-scl_error_t         graph_delete_vertex                     (graph_t * const __restrict__ gr, uint64_t vertex);
+scl_error_t         graph_delete_edge                       (const graph_t * const __restrict__ gr, size_t first_vertex, size_t second_vertex);
+scl_error_t         graph_delete_all_edges                  (const graph_t * const __restrict__ gr, size_t first_vertex, size_t second_vertex);
+scl_error_t         graph_delete_vertex                     (graph_t * const __restrict__ gr, size_t vertex);
 
 size_t              get_graph_size                          (const graph_t * const __restrict__ gr);
 
-size_t              graph_bfs_traverse                      (const graph_t * const __restrict__ gr, uint64_t start_vertex, uint64_t * __restrict__ vertex_path);
-size_t              graph_dfs_traverse                      (const graph_t * const __restrict__ gr, uint64_t start_vertex, uint64_t * __restrict__ vertex_path);
+size_t              graph_bfs_traverse                      (const graph_t * const __restrict__ gr, size_t start_vertex, size_t * __restrict__ vertex_path);
+size_t              graph_dfs_traverse                      (const graph_t * const __restrict__ gr, size_t start_vertex, size_t * __restrict__ vertex_path);
 
 uint8_t             graph_has_cycle                         (const graph_t * const __restrict__ gr);
-size_t              graph_vertex_past_vertices              (const graph_t * const __restrict__ gr, uint64_t start_vertex, uint64_t * __restrict__ vertex_path);
-size_t              graph_vertex_future_vertices            (const graph_t * const __restrict__ gr, uint64_t start_vertex, uint64_t * __restrict__ vertex_path);
-size_t              graph_vertex_anticone_vertices          (const graph_t * const __restrict__ gr, uint64_t start_vertex, uint64_t * __restrict__ vertex_path);
-size_t              graph_tips_vertices                     (const graph_t * const __restrict__ gr, uint64_t * __restrict__ vertex_path);
+size_t              graph_vertex_past_vertices              (const graph_t * const __restrict__ gr, size_t start_vertex, size_t * __restrict__ vertex_path);
+size_t              graph_vertex_future_vertices            (const graph_t * const __restrict__ gr, size_t start_vertex, size_t * __restrict__ vertex_path);
+size_t              graph_vertex_anticone_vertices          (const graph_t * const __restrict__ gr, size_t start_vertex, size_t * __restrict__ vertex_path);
+size_t              graph_tips_vertices                     (const graph_t * const __restrict__ gr, size_t * __restrict__ vertex_path);
 
-size_t              graph_topological_sort                  (const graph_t * const __restrict__ gr, uint64_t * __restrict__ vertex_path);
-scl_error_t         graph_dijkstra                          (const graph_t * const __restrict__ gr, uint64_t start_vertex, long double * __restrict__ vertex_dists, uint64_t * __restrict__ vertex_parents);
-scl_error_t         graph_prim                              (const graph_t * const __restrict__ gr, uint64_t start_vertex, long double * __restrict__ vertex_dists, uint64_t * __restrict__ vertex_parents);
+size_t              graph_topological_sort                  (const graph_t * const __restrict__ gr, size_t * __restrict__ vertex_path);
+scl_error_t         graph_dijkstra                          (const graph_t * const __restrict__ gr, size_t start_vertex, long double * __restrict__ vertex_dists, size_t * __restrict__ vertex_parents);
+scl_error_t         graph_prim                              (const graph_t * const __restrict__ gr, size_t start_vertex, long double * __restrict__ vertex_dists, size_t * __restrict__ vertex_parents);
 scl_error_t         graph_floyd_warshall                    (const graph_t * const __restrict__ gr, long double ** __restrict__ vertices_dists);
 
 uint8_t             graph_is_strongly_connected             (const graph_t * const __restrict__ gr);
-uint64_t**          graph_strongly_connected_components     (const graph_t * const __restrict__ gr, size_t *number_of_scc);
+size_t**            graph_strongly_connected_components     (const graph_t * const __restrict__ gr, size_t *number_of_scc);
 
 #endif /* GRAPH_UTILS_H_ */
